v1/requirements.txt
pydantic>=2.0
python-dotenv>=1.0
pytest>=7.0
zai